3.13 Support + Version Bump
