+++ .github/workflows/main.yml
@@ -6,7 +6,7 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        python-version: ["3.11", "3.12"]
+        python-version: ["3.11", "3.12", "3.13"]
         poetry-version: ["1.1.15", "1.6.1"]
         os: [macos-latest, windows-latest]
     runs-on: ${{ matrix.os }}
